backend/requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
psycopg2-binary>=2.9.9
djangorestframework-simplejwt>=5.3.0
django-otp>=1.3.0
qrcode>=7.4.2
celery>=5.3.4
redis>=5.0.1
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
Pillow>=10.1.0
cryptography>=41.0.7
python-decouple>=3.8
django-extensions>=3.2.3
drf-spectacular>=0.26.5
anthropic>=0.8.0
openai>=1.3.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.111.0
